test: replaced unittest test files with the new pytest files
